third_party/WebKit/LayoutTests/push_messaging/push-messaging-expected.txt

Tests for the Push API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS !!navigator.push is true
PASS !!navigator.push.register is true
PASS String(navigator.push.register("senderId")) is "[object Promise]"
PASS error.name is "AbortError"
PASS successfullyParsed is true

TEST COMPLETE
